Added and Changed TrendingAuction (read .docx), Added some countBots and countBidsInWindowTime

+++ src/main/java/com/auctionapp/auctionappjava/common/dto/AuctionTrendResponse.java
@@ -13,7 +13,7 @@ public record AuctionTrendResponse(
         AuctionStatus status,
         int bidderCount,
         int bidCount,
-        int trendScore,
+        BigDecimal trendScore,
         String trendLabel,
         String reason
 ) implements Serializable {
